Bump pre-release version to 1.3.0a0 from 1.2.2
Even if experimental, we are adding a (backward-compatible) functionality in
the remote streaming raster import.

According to SemVer, every minor upgrade must reset the patch.

Please note that since setuptools follows PEP 440, we must use the "a0"
suffix, which is anyway SemVer-compatible, to avoid normalization like

"UserWarning: Normalizing '1.3.0-a' to '1.3.0a0'"

see
https://packaging.python.org/en/latest/specifications/version-specifiers/#version-specifiers.