Refactor APIClient to introduce a separate client type
This introduces a BaseAPIClient base class that is inherited by
DetectorPlatformClient, which is the wrapper around Picterra's
Detector platform public api.

This prepares the ground for a PlotsAnalysisPlatformClient, which
will wrap Picterra's Plots Analysis platform public api.

Note that to preserve backward compatibility, the DetectorPlatformClient
is still exported as APIClient.