TracerClient: change order of arguments to allow future "url flattening"
We will likely simplify the URL structure so that getting a report doesn't
require passing the plots group and plots analysis ids, but only the report id.

By reversing argumenst to have the most specific id first, we will be able
to make this change in a backward-compatible manner, by just making the
other ids optional.